rfnoc: tx streamer: Default spp to multiple of CHDR width
Unless the user specifies an spp value in a Tx streamer by means of
a stream argument, the Tx streamer now holds additional logic that
aligns the spp value with the CHDR width. For example, in an X410 100
GbE image, the CHDR width is typically 512 bits, which corresponds to 16
samples (sc16), so spp will default to value that's a multiple of 16.

Previously, the default was to simply use the maximum MTU. However,
there have been circumstances observed where using a non-aligned spp
value causes suboptimal behaviour.

For example, when the CHDR width is 512 bits, and the discovered MTU is
8016 bytes, the default spp value would be 1988 samples per packet.
However, a value of 1984 samples per packet causes the CHDR packet to be
of a length that is a multiple of 64 bytes (== 512 bits) which can
improve performance.

This change does not affect applications where the spp size is already
manually configured, and has almost no influence on use cases where the
CHDR width is 64, because that just means the optimal number of samples
per packet should be even.

Under the hood, the following changes were made:
- The CHDR width is passed into the streamer as a reserved key
  (__chdr_width) from the infrastructure.
- After connecting the Tx streamer to the graph, when the MTU is known,
  we add a check for the spp value to make it an integer multiple of the
  CHDR width if not specifically given (like a "soft atomic item size").

Co-authored-by: Jörg Hofrichter <joerg.hofrichter@emerson.com>