cmake: Require NumPy component in FindPython3
In UHDPython.cmake, when calling find_package(Python3), we now add the
NumPy component. However, because this component was added in CMake 3.14,
and our minimum version is 3.12, we do this conditionally. For older
CMake versions, we rely on UHD_PYTHON_CHECK_MODULE_VERSION() to detect
numpy, which can fail to properly find the include headers in some
corner cases.