rfnoc: fft: Add support for multiple samples per clock
This adds support to the FFT block controller for the NIPC parameter.
Changing the NIPC parameter in the FPGA build allows the resulting
FFT block to process multiple items/samples per clock cycle, which
leads to higher throughput. As of compat version 3.1, the NIPC value
may be greater than 1 and must be a power of 2. The packet size and
cyclic prefix lengths used must also be a multiple of this NIPC
value.