image builder: grc import: Improve clock connection identification
In a GRC file, there is no difference between a clock connection and
any other connection. We add some code to improve the
auto-identification logic in the GRC import code.

This solves this particular problem: On X440, we have two clocks (radio0
and radio1) that share the same name with IO ports. If we draw
a connection from _device_:radio0 -> radio_block:radio, then it is
ambiguous if we are referring to a clock connection, or the IO port
connection.

In GRC we already append `_clk` to clock names to avoid the naming
ambiguity. In the GRC import code, we strip the `_clk` suffix. This
patch makes it so that we remember if we stripped off the suffix to help
identifying clock connections.