ci: streaming/hw tests: fix issue with increased FPGA compat num
The way the current CI implementation handles the artifacts has
an issue if the FPGA is taken from the FPGA pipeline and the FPGA
major compat number is increased. MPM in the filesystem already
expects the increased compat number but the installed FPGA image
under /lib/firmware still has the old compat number. As a result,
MPM is unable to initialize properly.

Fix this by deleting the installed FPGA image under /lib/firmware
and then restarting MPM.

Co-authored-by: Ahmed Salman <ahmed.salman@ni.com>