uhd: Move gps_ctrl.hpp into uhdlib
This is technically an API change, but it was never possible for any
user of UHD to instantiate a uhd::gps_ctrl without defining it
themselves. There is no public API that returns a `uhd::gps_ctrl`
object. It's strictly an internal object, and is thus moved outside of
the public API space.