cmake: Remove minimum CMake version from CMakeRC.cmake
This line is redundant (we call CMakeRC.cmake from other cmake files,
and we have our own cmake_minimum_required() call in
host/CMakeLists.txt), and it is also less strict than our own
requirements (we currently require at least 3.12 for all of UHD). With
recent CMake versions, this causes warnings so we remove this line.

This is the warning that is now removed:

```
CMake Deprecation Warning at cmake/Modules/CMakeRC.cmake:63
(cmake_minimum_required):
Compatibility with CMake < 3.5 will be removed from a future version of
CMake.
Update the VERSION argument <min> value or use a ...<max> suffix to tell
CMake that the project does not need compatibility with older versions.
```