host: dpdk: add error check for valid dpdk_lcore
This commit fixes the freezing behaviour of UHD
application running with DPDK. The Terminal
completely freezes when the dpdk_lcore value
mentioned in the uhd config file is not listed
in dpdk_corelist. This fix checks for this
condition and if not met, throws a runtime error.
This only affects when uhd is run with use_dpdk=1.