ci: Merge FPGA- with rest of pipeline
This changes makes the FPGA builds part of the monopipeline. A major
paradigm shift is the notion that the various hardware test pipelines no
longer have a parameter from where to download FPGA artifacts; they now
assume that artifacts are available in the current pipeline. That means
if we want to run a job like job-uhd-rf-tests.yml, we must make sure
that the FPGA artifacts are available, even if that means downloading
the artifacts from a different pipeline and re-uploading them. This has
the additional advantage that failing pipelines now will have all
artifacts directly associated with them.

The way we maintain dependencies correctly in the pipelines, without
having to build all FPGA bitfiles, is that there is a new step for every
bitfile, the "guarantee FPGA artifact step", which has both the
changeset analyzer and the build-fpga steps as a dependency. If the
build-fpga dependency is executed, the guarantee-artifacts step has
to wait until it is complete, but then is a no-op. We use a variable set
by the build-fpga step to identify successful completion of the bitfile
generation.
If the build-fpga dependency is skipped, then we immediately execute the
guarantee-artifacts step, which pulls bitfiles from earlier runs, and
makes them available as artifacts for the rest of the pipeline.

Jobs such as RF tests or streaming tests no longer have the option to
pull bitfiles from wherever. Instead, they assume that the artifacts are
available. An exception is the B200, because we currently don't build
those bitfiles as part of the pipeline.