host: Remove DMA FIFO id from pool
Requesting a new DMA data FIFO allocated an element in the
DMA FIFO pool which was never released when the DMA was release.
Thus recreating the FIFO exhausted the pool quickly. This commit
adds a callback to the RIO link which is used by the PCIe manager
to erase the FIFO id from the pool when the link goes out of scope.

Co-authored-by: Martin Braun <martin.braun@ettus.com>