ci: Manage all build-fpga-target flags in analyzer step
Instead of tracking the various "build_fpgaimg_*" flags everywhere, we
consume them in the changeset analyzer step. If those flags are provided,
they will set the appropriate flags in the UhdTestList, thus making all
those flags obsolete.

The behaviour of the pipeline is not affected, but it is simplified by
front-loading all settings into the changeset analyzer step.