rfnoc: Make e31x the default device for OOT modules
The E31x has the advantage that it can be targeted from the free tier of
Vivado, which means users can start writing blocks and running
testbenches without having to purchase a Vivado license. This is already
possible today by specifying -DUHD_FPGA_DEFAULT_DEVICE=e31x when running
cmake, but we think this will further lower the entry level for RFNoC
developers by making the free-tier option the default.