rfnoc_image_builder: Improve connection checking
When the image core YAML file contained problems such as:

- Invalid block names
- Invalid port names
- Incompatible connections of IO ports

...in a connection definition, the most likely outcome would be an
exception thrown with a cryptic message like this:

 AttributeError: 'CommentedMap' object has no attribute 'io_ports'

With this change, all connections are scanned for validity before
continuing. Also, to make the the error message outputs more useful, the
list of valid connections are folded into the individual error message,
and no longer towards the end of the image builder output.

For example, the error message could now look like this:

⛔   There are errors in the image core file's connection settings:
Destination port 'nullSink0.in_0' not found. Available ports on \
       'nullSink0': sink, loop
Source port 'nullSink0.out_0' not found. Available ports on \
       'nullSink0': source, loop