python: image_builder: Use path APIs instead of @CONFIG_PATH@
This removes a CMake-generated value for the package data path from the
image builder, and instead uses the get_pkg_data_path() API call which
is now always available.

This fixes the issue where if the package was built on a different
system than where it is run, the image builder can't find the RFNoC
input files.