examples: gr-rfnoc_gain: Add dependencies to CMake
This amends the CMake for gr-rfnoc_gain to search for and require UHD,
and, using the new rfnoc-gainConfig.cmake file, to search for and
require rfnoc-gain itself. While it is possible to create valid OOT
modules without this linkage, it is a sensible default and may be
required for certain features.