host: add retransmission capabilities to tx streamer
With this commit the host enables retranmit capabilities to tx
streamer. This can be used for host to FPGA transmission that
are real time depended but need absolut reliability. The main
use case will be transmissions from host to replay block. The
current implementation supports only single channel transmissions.
To enable retransmissions the streamer has to change the op mode
using the new STREAM_MODE_STOP_ON_SEQ_ERROR. The streamer will
check whether the FPGA reports that this mode was enabled before
proceeding. When enabled on the FPGA the FPGA will reject data
from the host when it detects a sequence error. Only after a new
initialize stream command is sent the host can transfer more data
to the FPGA. When the tx streamer detects a sequence error from the
FPGA it will stop the current send command and return the samples
sent so far. Note: This might be more than the FPGA has actually
processed. It is up to the user logic to determine if and where
the transmission has to restart (e.g. query the replay block for
arrived samples). If the tx streamer calls send again the
reinitialization is triggered automatically.