Host: Allow setting the path to the uhd.dll directly
On Windows allow the user to set the path to the uhd.dll so one
can load the Python bindings for UHD without actually installing it.
uhd.dll and libpyuhd.pyd are not build into the same folder and thus
Python will not find the uhd.dll automatically unless the Python
bindings are actually installed.

Co-authored-by: Martin Braun <martin.braun@ettus.com>