uhd: rfnoc: Move detail::graph_t into public API
This adds a publicly available header (uhd/rfnoc/detail/graph.hpp) which
has the same API as uhd::rfnoc::detail::graph_t before, but without any
private implementation details leaked.

The main reason for this change is to enable RFNoC OOT modules to
implement their own unit tests on block controllers. The majority of the
files changed are in order to change the include path or other
naming-related changes.

A note on host/lib/rfnoc/graph.cpp: The unit test for this file requires
access to non-public attributes and methods, so we add defines to
conditionally compile private symbols only.