uhd: Deprecate pimpl.hpp
This file promotes usage of std::shared_ptr for pimpl types, which is
not good practice. Also, it doesn't really do all that much: The macros
it defines are almost as long as their implementations.

Therefore, we remove its usage from UHD and replace it with standard
pointer types. For USRP1 and USRP2, usage of unique_ptr is not possible
without changing more of the implementation, so a shared_ptr is used.

The header itself is not removed, but the macros are declared
deprecated.