rfnoc_image_builder: Fix path selection for IP build dir
Inside the image builder, we would first sanitize the path for the IP
build dir (--build-ip-dir), but then we used the non-sanitized path to
generate the 'make' command. This meant that if --build-ip-dir was not
specifically provided on the command line, the default path was
incorrect. This would fail, e.g., if the source code is only available
as read-only.