host: zbx: Only update pwr in default gain profile
As ZBX uses a power calibration, it can be set to work with
power_reference settings. When retuning, it will automatically pick the
right gain values from the power calibration to provide approximately
the same power output/input as on the previous frequency. However when
switching to a different than the default gain profile, retuning will
lead to an error message as the power update method does only account
for the nameless gain setting which is only valid for the default gain
profile. Therefore, with this change we only run the update method
during retuning when we are in the default gain profile and warn
otherwise.