ci: Fix Device or resource busy bug in pipeline
This fixes the Device or resource busy bug in the x4xx hardware test
pipeline step. To enable different bitfiles than the one that were baked
into the file system image, we delete the installed one and install a
different one. This involves bringing up MPM without a bitfile. So when
the next bitfile gets installed, resources from the still running
bitfile cannot be cleaned up properly, like the mtdblocks of the X410
flash memories. To still clean that up, we now unmount these before we
start up MPM without a bitfile again. We ignore any errors that might
come up here as these steps are executed on devices without flash
memories on the daughtercards, too.