rfnoc_image_builder: Fix GRC import with empty default target
When GNU Radio Companion (GRC) stores files that represent RFNoC
bitfiles, it is possible that there is no "default target" provided. In
this case, the rfnoc_image_builder would incorrectly create an empty
default target, which would overwrite the actual build target (e.g., the
build target for an X410 should always be "X410", but it could be set to
empty-string).

In this fix, an empty default target is simply omitted from the
generated YAML file, and the image builder can insert its own default
target.