rfnoc_image_builder: Fix slicing notation
This fixes an issue where the slicing notation in YAML files was broken.
This is where you specify connections as such:

- { srcblk: "block1", srcport: "port1[7:0]",
    dstblk: "block2", dstport: "port2[7:0]" }

This commit also fixes an issue in the error reporting, where invalid
port names (that would not match a regex) should trigger an error line,
but this was not being printed out.