host: add helper routines to transfer data between host and replay (DRAM)
Transferring samples between host and replay block is a cumbersome job
with a lot of boiler plate code. This adds two new helper routines to
send data to the replay block receive data from the replay block into
some data buffer.
The interface is designed to be as close as possible to existing streamer
code in term of data types and names. Using Python this is just:

    tx_data = np.arange(num_samples, dtype=np.complex64)
    uhd.utils.dram.upload(graph, tx_data, "fc32")

    rx_data = np.zeros(num_samples, dtype=np.complex64)
    uhd.utils.dram.download(graph, rx_data, "fc32")