host: fix b200_image_loader to return correct eeprom and iface
get_b200_iface returns the correct device handle but it always
returns the eeprom and iface values from the last iteration of
the dev_handle for loop. This results into incorrect image loading
in case of multiple b200 are connected.
Fix is using the device handle to query eeprom and iface again
before returning.