host: rework the download of data from replay
Currently we download all data at once from replay in the DRAM utils.
This gets problematic on sequencing error on connections with multiple
hops for the UDP transport. In that case we need to query the replay
block for status but because of the sequence errors the response can
get lost and UHD has no measure to handle lost request responses.

The alternative implementation now downloads the data in chunks and
retriggers a complete chunk in case of an error. That way we do not
need to query the replay block for it's position.