lib: Avoid using braced-initializers for meta ranges
Because a meta_range_t is a vector of ranges, the following syntax is
ambiguous, if not to the compiler then to human readers:

```cpp
meta_range_t r{0.0, 1.0};
// Not intuitively clear if this means r{range_t(0.0), range_t(1.0)}, or
// if it means r{range_t(0.0, 1.0)}.
```

Within UHD, we modify some calls to meta_range_t constructors that are
using braced initializers to explicitly use other constructors.

Note that the ambiguous usage of the braced initializer syntax is
allowed and documented in ranges.hpp, we simply remove it to clarify the
intent of certain lines. Also, non-ambiguous usage of the braced
initializer ctor is also perfectly fine, e.g.:

```cpp
meta_range_t r{range_t(0.0), range_t(1.0), range_t(4.0)};
```