fpga: rfnoc: Fix alignment of first sample
Fixed an issue where the first sample could have a different alignment
from the rest of the samples. This could manifest as a missing or
duplicated sample in the data stream.

This was most likely to occur when doing a timed RX/TX followed by an
untimed RX/TX, in which case a sample might be missing at the
beginning of the burst.