x300: Enable `force_mtu` device argument
This device argument can be used where the automatic MTU discovery fails,
e.g., because of a particularly restrictive network configuration. Note
that it should be used with care, because UHD will use this value for
Ethernet frame sizes, even if the underlying network connection does not
support them.

If the desired goal is to override the MTU for a particular link, then
it may be more sensible to provide an `mtu` stream argument, or,
depending on the situation, a `recv_frame_size` or `send_frame_size`
stream argument instead. The main purpose of the `force_mtu` device
argument is to disable the automatic MTU discovery, and provide UHD with
a substitute value instead.

It may be used as follows:

```cpp
auto my_usrp =
  uhd::usrp::multi_usrp("type=x300,addr=192.168.30.2,force_mtu=8000");
// UHD will now assume that 8000 is a valid MTU, without further checks
```

Note that it remains possible to configure the Ethernet frame sizes
using the `recv_frame_size` or `send_frame_size` arguments, but those
will not disable the MTU discovery:

```cpp
auto my_usrp =
  uhd::usrp::multi_usrp(
    "type=x300,addr=192.168.30.2,recv_frame_size=8000");
// Path MTU may be smaller than 8000 if the USRP is not able to handle
// this MTU size
```