mpmd: Enable `force_mtu` device argument
This device argument can be used where the automatic MTU discovery fails,
e.g., because of a particularly restrictive network configuration. Note
that it should be used with care, because UHD will use this value for
Ethernet frame sizes, even if the underlying network connection does not
support them.

Also, for 1 GbE connections, the default MTU is always 1500 bytes,
regardless of the detected maximum MTU. This means that when attempting
to increase the MTU over a 1 GbE link, the recv_frame_size stream (or
device) argument also needs to be provided.

See the corresponding x300 commit for additional for information. This
enables the same feature for all MPM devices (X4xx, E3xx, N3xx).
Manual pages were also updated in this commit.