fpga: x400: Unset MGT LOC constraints before replacing them
Vivado by default will not allow you to place cells in locations where
cells have already been placed. Our IP is configured with a default
location and we use constraints to move the location when the
constraints are loaded. However, depending on the order in which
constraints are read, sometimes our builds try to place something into
a location that is occupied because the cell there has not yet been
moved.

This PR fixes this issue for the QSFP ports by adding constraints to
unset any existing LOC placement constraints for the associated MGT
before applying the new constraint. This ensures that the new
constraint gets applied and not rejected by Vivado.