fpga: x4xx_dio: remove all direction muxes
On X4XX, the DIO pins are driven by a driver chip which is controlled
by the MB CPLD. Therefore software side (=MPM) has to ensure that the
direction which is set by the FPGA is in sync with the direction that
is configured in the CPLD. This is given for the DIO_DIRECTION
register. Using other sources would be unsave because it could happen
that there are 2 drivers on the GPIO line. Consequently, use the
DIO_DIRECTION register in all cases and do not use the other
direction sources (PS, ATR states, digital interface).