rfnoc: Allow asynchronous action handling
This provides a new function argument to post_action() to allow actions
to be executed asynchronously. For backward compatibility reasons, the
default value is "synchronous", although for many cases, an asynchronous
execution is a better choice.

This is implemented by adding a dedicated action handler thread (there
is one such thread per graph). All actions are handled, sequentially,
inside this thread. For the synchronous mode, there is a synchronization
point. This is also how exceptions from the action handler thread are
propagated into the caller thread.